+++ blocks/columns/columns.css
@@ -33,7 +33,7 @@
 hr.column-underline {
     margin: 0;
     border-bottom-width: 3px;
-  }
+}
 
 .columns.promo .columns-promo-text-wrapper {
     order: 2;
